requirements.txt
asgiref==3.11.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.4
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
markdown2==2.5.4
PyJWT==2.11.0
requests==2.32.5
soupsieve==2.7
sqlparse==0.5.5
typing_extensions==4.14.1
urllib3==2.6.3
